move getFormattedEventDates into helpers
